requirements.txt
fastapi==0.110.1
httpx==0.27.0
python-dotenv==1.0.1
uvicorn==0.29.0